src/core/en_azure/requirements.txt
openai==1.6.0
streamlit==1.30.0
python-dotenv==1.0.1
tiktoken==0.5.2